Onzeit/backend/requirements.txt
Django==1.9.8
MySQL-python==1.2.5
Pygments==2.1.3
argparse==1.2.1
django-cors-headers==1.1.0
djangorestframework==3.4.0
httpie==0.9.4
httplib2==0.9.2
pycrypto==2.6.1
pytz==2016.6.1
requests==2.10.0
six==1.10.0
twilio==5.4.0
var-dump==1.0.1
wsgiref==0.1.2
